requirements.txt
deepgram-sdkdeepgram-sdk

openaiopenai

python-dotenvpython-dotenv

pydubpydub

faster-whisperfaster-whisper

torchtorch

torchaudiotorchaudio

tqdm>=4.65.0tqdm>=4.65.0

whisperxwhisperx

googlemapsgooglemaps

psutilpsutil

flask
flask-socketio
pandas
requests
sqlite3
pathlib